// week15_final/src/types/tmdb.ts
export interface Genre {
    id: number;
    name: string;
}

export interface Video {
    iso_639_1: string;
    iso_3166_1: string;
    name: string;
    key: string;
    site: string;
    size: number;
    type: string;
    official: boolean;
    published_at: string;
    id: string;
}

export interface VideoResponse {
    id: number;
    results: Video[];
}

export interface Image {
    aspect_ratio: number;
    height: number;
    iso_639_1: string | null;
    file_path: string;
    vote_average: number;
    vote_count: number;
    width: number;
}

export interface ImageResponse {
    backdrops: Image[];
    logos: Image[];
    posters: Image[];
}
// ...
export interface Credits {
    cast: CastMember[];
    crew: CrewMember[];
}

export interface Review {
    author: string;
    author_details: {
        name: string;
        username: string;
        avatar_path: string | null;
        rating: number | null;
    };
    content: string;
    created_at: string;
    id: string;
    updated_at: string;
    url: string;
}

export interface ReviewResponse {
    page: number;
    results: Review[];
    total_pages: number;
    total_results: number;
}

export interface Keyword {
    id: number;
    name: string;
}

export interface KeywordResponse {
    keywords: Keyword[];
}

export interface ProductionCompany {
    id: number;
    logo_path: string | null;
    name: string;
    origin_country: string;
}

export interface CompanyDetails extends ProductionCompany {
    description?: string;
    headquarters?: string;
    homepage?: string;
    movies?: MovieListResponse;
}

export interface ProductionCountry {
    iso_3166_1: string;
    name: string;
}

export interface SpokenLanguage {
    english_name: string;
    iso_639_1: string;
    name: string;
}

export interface Movie {
    adult: boolean;
    backdrop_path: string | null;
    belongs_to_collection: null | object;
    budget: number;
    genres: Genre[];
    genre_ids?: number[];
    homepage: string | null;
    id: number;
    imdb_id: string | null;
    original_language: string;
    original_title: string;
    overview: string;
    popularity: number;
    poster_path: string | null;
    production_companies: ProductionCompany[];
    production_countries: ProductionCountry[];
    release_date: string;
    revenue: number;
    runtime: number | null;
    spoken_languages: SpokenLanguage[];
    status: string;
    tagline: string | null;
    title: string;
    video: boolean;
    vote_average: number;
    vote_count: number;
    // Extended properties often included in our app via append_to_response
    videos?: VideoResponse;
    credits?: Credits;
    images?: ImageResponse;
    reviews?: ReviewResponse;
    keywords?: KeywordResponse;
    similar?: MovieListResponse;
}

export interface MovieListResponse {
    page: number;
    results: Movie[];
    total_pages: number;
    total_results: number;
}
// ...
export interface DailyMovie extends Movie {
    source?: string;
    recommendationContext?: {
        name: string;
        description: string;
        label?: string;
    };
}